# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "API documentation for CRM system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/individuals/{id}": {
            "get": {
                "tags": [
                    "Individuals"
                ],
                "summary": "Получить физическое лицо",
                "description": "Получает данные физического лица по ID",
                "operationId": "805a4615d1a3144e8fba1dd5e91bf0dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID физического лица",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные физического лица",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Иван"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Иванов"
                                                },
                                                "middle_name": {
                                                    "type": "string",
                                                    "example": "Иванович"
                                                },
                                                "full_name": {
                                                    "type": "string",
                                                    "example": "Иванов Иван Иванович"
                                                },
                                                "short_name": {
                                                    "type": "string",
                                                    "example": "Иванов И.И."
                                                },
                                                "status_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "position_id": {
                                                    "type": "integer",
                                                    "example": 1,
                                                    "nullable": true
                                                },
                                                "login": {
                                                    "type": "string",
                                                    "example": "ivan123",
                                                    "nullable": true
                                                },
                                                "is_company_employee": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "creator_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-08-08T13:47:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Физическое лицо не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Физическое лицо не найдено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}